细节优化

+++ App/Database/DBFolder.swift
@@ -23,6 +23,11 @@ class DBFolder: ObservableObject {
     
     func deleteFile(_ audio: Audio) throws {
         os_log("\(Logger.isMain)\(self.label)删除 \(audio.url)")
+        
+        if fileManager.fileExists(atPath: audio.url.path) == false {
+            return
+        }
+        
         try fileManager.removeItem(at: audio.url)
     }
